+++ internal/app/loki-multi-tenant-proxy/server.go
@@ -2,32 +2,76 @@ package proxy
 
 import (
 	"fmt"
-	"log"
 	"net/http"
 	"net/http/httputil"
 	"net/url"
 
 	"github.com/giantswarm/loki-multi-tenant-proxy/internal/pkg"
 	"github.com/urfave/cli/v2"
+	"go.uber.org/zap"
 )
 
 // Serve serves
 func Serve(c *cli.Context) error {
 	lokiServerURL, _ := url.Parse(c.String("loki-server"))
-	serveAt := fmt.Sprintf(":%d", c.Int("port"))
+	addr := fmt.Sprintf(":%d", c.Int("port"))
 	authConfigLocation := c.String("auth-config")
 	authConfig, _ := pkg.ParseConfig(&authConfigLocation)
 	authConfig.KeepOrgID = c.Bool("keep-orgid")
 
-	http.HandleFunc("/", createHandler(lokiServerURL, authConfig))
-	if err := http.ListenAndServe(serveAt, nil); err != nil {
-		log.Fatalf("Loki multi tenant proxy can not start %v", err)
-		return err
+	logLevel := c.String("log-level")
+	if logLevel == "" {
+		logLevel = "INFO"
 	}
-	return nil
-}
 
-func createHandler(lokiServerURL *url.URL, authConfig *pkg.Authn) http.HandlerFunc {
-	reverseProxy := httputil.NewSingleHostReverseProxy(lokiServerURL)
-	return LogRequest(BasicAuth(ReverseLoki(reverseProxy, lokiServerURL), authConfig))
+	var logger *zap.Logger
+	{
+		zapConfig := zap.NewProductionConfig()
+		level, err := zap.ParseAtomicLevel(logLevel)
+		if err != nil {
+			return cli.Exit(fmt.Sprintf("Could not parse log level %v", err), -1)
+		}
+		zapConfig.Level = level
+
+		logger = zap.Must(zapConfig.Build())
+		defer logger.Sync()
+	}
+
+	errorLogger, err := zap.NewStdLogAt(logger, zap.ErrorLevel)
+	if err != nil {
+		return cli.Exit(fmt.Sprintf("Could not create standard logger %v", err), -1)
+	}
+
+	var reverseProxy *httputil.ReverseProxy
+	{
+		reverseProxy = &httputil.ReverseProxy{
+			Rewrite: func(r *httputil.ProxyRequest) {
+				r.SetURL(lokiServerURL)
+				r.Out.Host = lokiServerURL.Host
+				r.Out.Header.Set("X-Forwarded-Host", lokiServerURL.Host)
+				orgID := r.In.Context().Value(OrgIDKey)
+
+				if orgID != "" {
+					r.Out.Header.Set("X-Scope-OrgID", orgID.(string))
+				}
+			},
+			ErrorLog: errorLogger,
+		}
+	}
+
+	handlers := Logger(
+		BasicAuth(
+			ReverseLoki(reverseProxy),
+			authConfig,
+		),
+		logger,
+	)
+
+	http.HandleFunc("/", handlers)
+	server := &http.Server{Addr: addr, ErrorLog: errorLogger}
+	if err := server.ListenAndServe(); err != nil {
+		return cli.Exit(fmt.Sprintf("Loki multi tenant proxy could not start %v", err), -1)
+	}
+	logger.Info("Starting HTTP server", zap.String("addr", addr))
+	return nil
 }